Renaming models folder as model by PR comment.
Issue: IBP-6185
Review: none
