+++ src/main/jhipster/src/main/webapp/app/site-admin/model/UserSearchRequest.ts
