+++ src/main/jhipster/src/main/webapp/app/site-admin/model/permission.model.ts
