+++ src/main/jhipster/src/main/webapp/app/site-admin/model/role-type.model.ts
