+++ src/main/jhipster/src/main/webapp/app/site-admin/model/role.model.ts
