+++ src/main/jhipster/src/main/webapp/app/site-admin/site-admin-context.ts
@@ -1,7 +1,7 @@
 import { Injectable } from '@angular/core';
 import { User } from '../shared/user/model/user.model';
-import { RoleType } from './models/role-type.model';
-import { Role } from './models/role.model';
+import { RoleType } from './model/role-type.model';
+import { Role } from './model/role.model';
 
 @Injectable()
 export class SiteAdminContext {
