+++ src/main/jhipster/src/main/webapp/app/site-admin/tabs/users/users-role-dialog.component.ts
@@ -4,11 +4,11 @@ import { UserEditDialogComponent } from './user-edit-dialog.component';
 import { UserService } from '../../services/user.service';
 import { RoleService } from '../../services/role.service';
 import { ActivatedRoute, Router } from '@angular/router';
-import { RoleType } from '../../models/role-type.model';
-import { Role } from '../../models/role.model';
+import { RoleType } from '../../model/role-type.model';
+import { Role } from '../../model/role.model';
 import { Crop } from '../../../shared/model/crop.model';
 import { Program } from '../../../shared/user/model/program.model';
-import { RoleFilter } from '../../models/role-filter.model';
+import { RoleFilter } from '../../model/role-filter.model';
 import { UserRole } from '../../../shared/user/model/user-role.model';
 import { User } from '../../../shared/user/model/user.model';
 import { SiteAdminContext } from '../../site-admin-context';
